refactor : model 패키지 내용 domain으로 이동
