+++ src/test/kotlin/lotto/domain/LottoCompanyTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.model.LottoTicket
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
