+++ src/test/kotlin/lotto/domain/YieldCalculatorTest.kt
@@ -1,8 +1,5 @@
 package lotto.domain
 
-import lotto.model.LottoResult
-import lotto.model.LottoResults
-import lotto.model.Prize
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
